requirements.txt
pytest==7.1.3
pytest-xdist==2.5.0
selenium==4.5.0
webdriver-manager==3.8.3
pytest-rerunfailures==10.2
pymysql==1.0.2
sqlalchemy==1.4.43
cryptography==38.0.4
allure-pytest==2.11.1
